pyproject.toml
[tool.poetry]
name = "matter"
version = "0.0.1"
description = ""
authors = ["Sergey Reshetnikov <resh.sersh@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.0"
pampy = "^0.3.0"
aiofiles = "^0.4.0"
attrs = "^19.3"
antlr4-python3-runtime = "^4.7"
uvloop = "^0.14.0"
starlette = "^0.13.2"

[tool.poetry.dev-dependencies]
pytest = "^5.3"
mypy = "^0.750.0"
pylint = "^2.4"
sphinx = "^2.2"
pytest-cov = "^2.8"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
